Don't merge leaf subranges with non-leaf ones
As we are generating line info only for leaf ranges, merging a leaf
range with a non-leaf one results in not generating line info for the
former (previously leaf) range.

(cherry picked from commit aadeac97667d3c369b3358b682902cb515dae2d2)

+++ substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/debugentry/Range.java
@@ -306,6 +306,13 @@ public class Range {
             /* cannot merge callers with different line numbers, move on. */
             return sibling;
         }
+        if (isLeaf() != sibling.isLeaf()) {
+            /*
+             * cannot merge leafs with non-leafs as that results in them becoming non-leafs and not
+             * getting proper line info
+             */
+            return sibling;
+        }
         /* splice out the sibling from the chain and update this one to include it. */
         unlink(debugContext, sibling);
         /* relocate the siblings children to this node. */
